+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - (x/evm/keeper) add method `SetCustomPrecompiles`
 - (precompile/common) add `GetMethodIDByName` method
+- (precompile/common) add `GetCosmosAddressFromEVMAddressArg` method
 
 ### Improvements